WIP: 新增 DomainEvent Extension for Test

+++ Domain/Events/ThroughStartEvent.cs
@@ -2,4 +2,4 @@
 
 namespace Domain.Events;
 
-public record ThroughStartEvent(string PlayerId, int GainMoney, decimal TotalMoney) : DomainEvent;
+public record ThroughStartEvent(string PlayerId, decimal GainMoney, decimal TotalMoney) : DomainEvent;
